raspberrypi/requirements.txt
paho-mqtt
opencv-python
influxdb-client
adafruit-circuitpython-dht
RPi.GPIO
gpiozero
